split gui.spell namespace in half. starting to prep a new item with inventory
